Move entrypoints to `impl`.
