+++ src/main/java/io/github/kvverti/bannerpp/impl/Bannerpp.java
@@ -1,4 +1,4 @@
-package io.github.kvverti.bannerpp;
+package io.github.kvverti.bannerpp.impl;
 
 import io.github.kvverti.bannerpp.api.LoomPatterns;
